installer/py2.7/pip_requirements.txt
## Installing offline pip requirements  (isf-py27)
## pip install -r isf2-pip-requirements.txt --no-index --find-links wheelhouse
## NOTE1: pip dependencies need to get installed before conda dependencies
##        in order for conda dependecies resolution to succeed. 
##        Alternatively installing conda dependencies without dependency checking.
## NOTE2: setuptools-scm needs to be installed in advance for the installation to succeed in Python 2.7
## pip install setuptools-scm
#pip==20.3.4
#alabaster==0.7.12
#Babel==2.9.1
#backports-abc==0.5
#backports.functools-lru-cache==1.6.4
#backports.shutil-get-terminal-size==1.0.0
#backports.ssl-match-hostname==3.5.0.1
#bitarray==0.8.1
#bleach==3.3.1
#blosc==1.5.1
bluepyopt==1.8.5
bokeh==0.13.0
#configparser==4.0.2
deap==1.2.2
#defusedxml==0.7.1
django-tagging==0.4.6
Django==1.8.19
#docutils==0.18.1
#efel==4.0.4
#entrypoints==0.3
#et-xmlfile==1.0.1
fasteners==0.14.1
future==0.16.0
# greenlet==1.1.2
httplib2==0.11.3
Jinja2==2.10
#imagesize==1.3.0
#ipykernel==4.10.1
#ipyparallel==6.2.5
#ipython-genutils==0.2.0
#ipython==5.10.0
#jupyter-client==5.3.5
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
# jupyter-core==4.3.0
#jupyter-core==4.6.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
kazoo==2.8.0
#lxml==4.6.4
#mistune==0.8.4
monotonic==1.5
#nbconvert==5.6.1
neo==0.7.1
#notebook==5.7.13
pandas-msgpack==0.1.4
#pandocfilters==1.5.0
parameters==0.2.1
#pathlib2==2.3.6
pathlib==1.0.1
#pexpect==4.8.0
#pickleshare==0.7.5
#prometheus-client==0.12.0
#prompt-toolkit==1.0.18
#ptyprocess==0.7.0
#Pygments==2.5.2
PyPDF2==1.26.0
plotly==4.14.3
pyparsing==2.2.0
#pyzmq==19.0.2
quantities==0.12.2
#redis==3.5.3
scandir==1.10.0
#scoop==0.7.1.1
#Send2Trash==1.8.0
#simplegeneric==0.8.1
SimpleITK==1.2.0
six==1.12.0
#snowballstemmer==2.2.0
#Sphinx==1.4.6
Sumatra==0.7.4
# tables==3.3.0 (provided via conda package pytables)
tenacity==7.0.0
#terminado==0.8.3
#testpath==0.4.4
#tqdm==4.62.3
#wcwidth==0.2.5
#webencodings==0.5.1
pytest==2.9.2